refactor(sanity): add reusable perspective navigator
